added registration dto with more fields

+++ playlist_generator/src/main/java/track_ninja/playlist_generator/models/Authority.java
@@ -30,7 +30,7 @@ public class Authority implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = AUTHORITY_ID)
-    private long authorityId;
+    private int authorityId;
 
     @Column(name = ROLE_TYPE)
     @NotNull
